Crate memo_rust

A library for processing text in the memos in Organizator.

As Organizator allows encrypting data on the client, this library will be compiled to webassembly and mostly used in the browser.

Encryption

The parts of the text that are supposed to be encrypted will be marked with Japanese quotes: 「secret」.
The encryption algorithm is AES-256 with a timestamp nonce and a counter. After encryption, the text will be base64 encoded.

Decryption

The text will be decrypted with the same algorithm and the same key. The bits to be decrypted are automatically detected by looking for base64-encoded sequences.

Presentation

Because long base64-encoded strings are not very readable, those sequences will be replaced with a shorter version followed by … (ellipsis).

Modules

aes

Functions

memo_decrypt

Decrypts the base64 encoded segments in a memo

memo_encrypt

Encrypts all segments surrounded by Japanes quotation marks in the memo 「secret」

prepare_memo_for_view

Truncates the large segments of base64 and adds an ellipsis inside input string (uses unsafe)

truncate_base64

Truncates the large segments of base64 if they are larger than max_b64 and adds an ellipsis.